Enhancement/70 velocity distribution plot

+++ src/lib/gui/models/SetupModel.cpp
@@ -25,9 +25,9 @@ void SetupModel::setTimeScale(double timeScale)
     simulationConfig_.simulationTimeScale = timeScale;
 }
 
-void SetupModel::setMostProbableSpeed(int mostProbableSpeed)
+void SetupModel::setMostProbableSpeed(double mostProbableSpeed)
 {
-    simulationConfig_.mostProbableSpeed = static_cast<double>(mostProbableSpeed);
+    simulationConfig_.mostProbableSpeed = mostProbableSpeed;
 }
 
 void SetupModel::setFPS(int FPS)
